+++ src/components/SidePanel/Tracing/StartStopTrace.tsx
@@ -11,7 +11,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import playSvg from '../../../../resources/play-circle.svg';
 import stopSvg from '../../../../resources/stop-circle.svg';
 import { startTrace, stopTrace } from '../../../features/tracing/nrfml';
-import { TraceFormat } from '../../../features/tracing/traceFormat';
+import { TraceFormat } from '../../../features/tracing/sinks';
 import { getIsTracing, getTaskId } from '../../../features/tracing/traceSlice';
 
 type StartStopProps = {