+++ src/components/SidePanel/__tests__/TraceCollector.test.tsx
@@ -5,6 +5,7 @@
  */
 
 import React from 'react';
+import nrfml from '@nordicsemiconductor/nrf-monitor-lib-js';
 
 import {
     setAvailableSerialPorts,
@@ -15,7 +16,10 @@ import {
     mockedCheckDiskSpace,
     render,
 } from '../../../utils/testUtils';
+import * as wireshark from '../../../utils/wireshark';
 import TraceCollector from '../Tracing/TraceCollector';
+
+jest.mock('../../../utils/wireshark');
 
 mockedCheckDiskSpace.mockImplementation(
     () =>
@@ -52,5 +56,75 @@ describe('TraceCollector', () => {
         const stopButton = await screen.findByText('Stop tracing');
         fireEvent.click(stopButton);
         expect(await screen.findByText('Start tracing')).toBeInTheDocument();
+    });
+
+    describe('wireshark installation', () => {
+        beforeEach(() => {
+            jest.resetAllMocks();
+        });
+
+        it('should display warning if wireshark is not installed and live tracing is selected', async () => {
+            jest.spyOn(wireshark, 'findWireshark').mockReturnValue(null);
+            const screen = render(<TraceCollector />, [...serialPortActions]);
+            const traceFormatButton = await screen.findByText('live');
+            fireEvent.click(traceFormatButton);
+            expect(
+                await screen.findByText('Wireshark not detected')
+            ).toBeInTheDocument();
+        });
+
+        it('should not display warning if wireshark is installed and live tracing is selected', async () => {
+            jest.spyOn(wireshark, 'findWireshark').mockReturnValue(
+                'path/to/wireshark'
+            );
+            const screen = render(<TraceCollector />, [...serialPortActions]);
+            const traceFormatButton = await screen.findByText('live');
+            fireEvent.click(traceFormatButton);
+            expect(
+                await screen.queryByText('Wireshark not detected')
+            ).not.toBeInTheDocument();
+        });
+    });
+
+    describe('sink configurations', () => {
+        beforeEach(() => {
+            jest.resetAllMocks();
+        });
+
+        it('should call nrfml start with selected sink configurations as arguments', async () => {
+            const screen = render(<TraceCollector />, serialPortActions);
+            fireEvent.click(await screen.findByText('raw'));
+            fireEvent.click(screen.getByText('Start tracing'));
+
+            expect(nrfml.start).toHaveBeenCalled();
+            // @ts-ignore -- ts doesn't know that nrfml.start has been mocked
+            const args = nrfml.start.mock.calls[0][0];
+            expect(args.sinks.length).toBe(1);
+        });
+
+        it('should call nrfml start with selected sink configurations as arguments', async () => {
+            const screen = render(<TraceCollector />, serialPortActions);
+            fireEvent.click(await screen.findByText('raw'));
+            fireEvent.click(await screen.findByText('pcap'));
+            fireEvent.click(screen.getByText('Start tracing'));
+
+            expect(nrfml.start).toHaveBeenCalled();
+            // @ts-ignore -- ts doesn't know that nrfml.start has been mocked
+            const args = nrfml.start.mock.calls[0][0];
+            expect(args.sinks.length).toBe(2);
+        });
+
+        it('should call nrfml start with selected sink configurations as arguments', async () => {
+            const screen = render(<TraceCollector />, serialPortActions);
+            fireEvent.click(await screen.findByText('raw'));
+            fireEvent.click(await screen.findByText('pcap'));
+            fireEvent.click(await screen.findByText('live'));
+            fireEvent.click(screen.getByText('Start tracing'));
+
+            expect(nrfml.start).toHaveBeenCalled();
+            // @ts-ignore -- ts doesn't know that nrfml.start has been mocked
+            const args = nrfml.start.mock.calls[0][0];
+            expect(args.sinks.length).toBe(3);
+        });
     });
 });