+++ src/features/tracing/nrfml.ts
@@ -5,112 +5,29 @@
  */
 
 import nrfml, { getPluginsDir } from '@nordicsemiconductor/nrf-monitor-lib-js';
-// eslint-disable-next-line import/no-unresolved -- Because this is a pure typescript type import which eslint does not understand correctly yet. This can be removed either when we start to use eslint-import-resolver-typescript in shared of expose this type in a better way from nrf-monitor-lib-js
-import { InsightInitParameters } from '@nordicsemiconductor/nrf-monitor-lib-js/config/configuration';
 import path from 'path';
-import { Device, getAppDataDir, logger } from 'pc-nrfconnect-shared';
-import { pathToFileURL } from 'url';
+import { getAppDataDir, logger } from 'pc-nrfconnect-shared';
 
-import { deviceInfo, selectedDevice } from '../../shouldBeInShared';
+import { selectedDevice } from '../../shouldBeInShared';
 import { TAction } from '../../thunk';
-import { autoDetectDbRootFolder } from '../../utils/store';
-import { fileExtension, sinkName, TraceFormat } from './traceFormat';
+import {
+    fileExtension,
+    requiresTraceDb,
+    sinkConfig,
+    TraceFormat,
+} from './sinks';
+import { detectModemFwUuid, detectTraceDB, sourceConfig } from './sources';
 import {
     getManualDbFilePath,
     getSerialPort,
+    getWiresharkPath,
     setDetectingTraceDb,
     setTaskId,
     setTraceData,
     TraceData,
 } from './traceSlice';
 
-const { displayName: appName } = require('../../../package.json');
-
 export type TaskId = number;
-const BUFFER_SIZE = 1;
-const CHUNK_SIZE = 256;
-
-const autoDetectDbCacheDirectory = path.join(getAppDataDir(), 'trace_db_cache');
-
-const autoDetectDbRootURL = pathToFileURL(autoDetectDbRootFolder).toString();
-
-const sourceConfig = (
-    manualDbFilePath: string | undefined,
-    useTraceDB: boolean,
-    additionalInitParameters: Partial<InsightInitParameters['init_parameters']>
-) => {
-    const initParameterForTraceDb =
-        manualDbFilePath != null
-            ? { db_file_path: manualDbFilePath }
-            : {
-                  auto_detect_db_config: {
-                      cache_directory: autoDetectDbCacheDirectory,
-                      root: autoDetectDbRootURL,
-                      update_cache: true,
-                      // eslint-disable-next-line no-template-curly-in-string -- Because this is no template string but the syntax used by nrf-monitor-lib
-                      trace_db_locations: ['${root}/config.json'] as unknown[],
-                  },
-              };
-
-    return {
-        name: 'nrfml-insight-source',
-        init_parameters: {
-            ...additionalInitParameters,
-            ...(useTraceDB ? initParameterForTraceDb : {}),
-            chunk_size: CHUNK_SIZE,
-        },
-        config: {
-            buffer_size: BUFFER_SIZE,
-        },
-    } as const;
-};
-
-const describeDevice = (device: Device) =>
-    `${deviceInfo(device).name ?? 'unknown'} ${device?.boardVersion}`;
-
-const additionalPcapProperties = (format: TraceFormat, device?: Device) => {
-    if (format === 'raw') return {};
-
-    return {
-        os_name: process.platform,
-        application_name: appName,
-        hw_name: device != null ? describeDevice(device) : undefined,
-    };
-};
-
-const sinkConfig = (format: TraceFormat, filePath: string, device?: Device) =>
-    ({
-        name: sinkName(format),
-        init_parameters: {
-            file_path: filePath,
-            ...additionalPcapProperties(format, device),
-        },
-    } as const);
-
-function detectTraceDB(progress: nrfml.Progress, detectedTraceDB: unknown) {
-    if (
-        progress.meta?.modem_db_path != null &&
-        detectedTraceDB !== progress.meta?.modem_db_path
-    ) {
-        detectedTraceDB = progress.meta?.modem_db_path;
-        logger.info(`Using trace DB ${detectedTraceDB}`);
-    }
-    return detectedTraceDB;
-}
-
-function detectModemFwUuid(
-    progress: nrfml.Progress,
-    detectedModemFwUuid: unknown
-) {
-    if (
-        progress.meta?.modem_db_uuid != null &&
-        detectedModemFwUuid !== progress.meta?.modem_db_uuid
-    ) {
-        detectedModemFwUuid = progress.meta?.modem_db_uuid;
-        logger.info(`Detected modem firmware with UUID ${detectedModemFwUuid}`);
-    }
-    return detectedModemFwUuid;
-}
 
 const convertTraceFile =
     (sourcePath: string): TAction =>
@@ -133,7 +50,7 @@ const convertTraceFile =
         const taskId = nrfml.start(
             {
                 config: { plugins_directory: getPluginsDir() },
-                sinks: [sinkConfig(traceFormat, destinationPath)],
+                sinks: [sinkConfig[traceFormat](destinationPath)],
                 sources: [
                     sourceConfig(manualDbFilePath, true, {
                         file_path: sourcePath,
@@ -186,20 +103,28 @@ const startTrace =
         const device = selectedDevice(getState());
         const filename = `trace-${new Date().toISOString().replace(/:/g, '-')}`;
         const traceData: TraceData[] = [];
+        let wiresharkPath: string | null;
+        let filePath = '';
         const sinkConfigs = traceFormats.map(format => {
-            const filePath =
-                path.join(getAppDataDir(), filename) + fileExtension(format);
+            if (format === 'live') {
+                wiresharkPath = getWiresharkPath(getState());
+            } else {
+                filePath =
+                    path.join(getAppDataDir(), filename) +
+                    fileExtension(format);
+            }
+
             traceData.push({
                 format,
                 path: filePath,
                 size: 0,
             });
-            return sinkConfig(format, filePath, device);
+            return sinkConfig[format](filePath, device, wiresharkPath);
         });
 
         const manualDbFilePath = getManualDbFilePath(getState());
 
-        if (!manualDbFilePath && traceFormats.includes('pcap')) {
+        if (!manualDbFilePath && requiresTraceDb(traceFormats)) {
             dispatch(setDetectingTraceDb(true));
         }
 
@@ -214,7 +139,7 @@ const startTrace =
                 sources: [
                     sourceConfig(
                         manualDbFilePath,
-                        traceFormats.includes('pcap'),
+                        requiresTraceDb(traceFormats),
                         {
                             serialport: { path: serialPort },
                         }
@@ -227,6 +152,10 @@ const startTrace =
                     logger.debug(`Full error: ${JSON.stringify(err)}`);
                 } else {
                     logger.info('Finished tracefile');
+                }
+                // stop tracing if Completed callback is called and we are only doing live tracing
+                if (traceFormats.length === 1 && traceFormats[0] === 'live') {
+                    dispatch(stopTrace(taskId));
                 }
             },
             progress => {