+++ src/utils/store.tsx
@@ -8,10 +8,7 @@ import { getPluginsDir } from '@nordicsemiconductor/nrf-monitor-lib-js';
 import path from 'path';
 import { getPersistentStore as store } from 'pc-nrfconnect-shared';
 
-import {
-    ALL_TRACE_FORMATS,
-    TraceFormat,
-} from '../features/tracing/traceFormat';
+import { ALL_TRACE_FORMATS, TraceFormat } from '../features/tracing/sinks';
 
 interface DevicePort {
     [serialNumber: string]: string;
@@ -21,14 +18,12 @@ const MANUAL_DB_FILE_PATH_KEY = 'dbFilePath';
 const WIRESHARK_EXECUTABLE_PATH_KEY = 'wiresharkExecutablePath';
 const TRACE_FORMATS = 'traceFormats';
 const SERIALPORTS = 'serialPorts';
-const HIDE_NRF_CMD_LINE_ALERT = 'hideNrfCmdLineAlert';
 
 interface StoreSchema {
     [MANUAL_DB_FILE_PATH_KEY]: string;
     [WIRESHARK_EXECUTABLE_PATH_KEY]: string | null;
     [TRACE_FORMATS]: TraceFormat[];
     [SERIALPORTS]: DevicePort;
-    [HIDE_NRF_CMD_LINE_ALERT]: boolean;
 }
 
 const AUTO_DETECT_DB_ROOT_RELATIVE_TO_PLUGINS_DIR = [
@@ -70,8 +65,3 @@ export const setSerialPort = (serialNumber: string, port: string) =>
         ...serialPorts(),
         [serialNumber]: port,
     });
-
-export const getHideNrfCommandLineAlert = () =>
-    store<StoreSchema>().get(HIDE_NRF_CMD_LINE_ALERT, false);
-export const persistHideNrfCommandLineAlert = () =>
-    store<StoreSchema>().set(HIDE_NRF_CMD_LINE_ALERT, true);